+++ test/common/cpu_info.cpp
@@ -267,6 +267,13 @@ bool cpu_has_sve() {
     return CpuInfo::current().has_sve;
 }
 
+bool cpu_has_sve_vl256() {
+    if (CpuInfo::current().has_sve) {
+        return (kai_get_sve_vector_length_u8() == 32);
+    }
+    return false;
+}
+
 bool cpu_has_sve2() {
     return CpuInfo::current().has_sve2;
 }